# pipeline/api-specs/openapi.yaml
components:
  schemas:
...
    CancelTaskRequest:
      additionalProperties: false
      properties:
        command:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Command
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - workspace_id
      - user_id
      title: CancelTaskRequest
      type: object
...
    DocumentCleanupResponse:
      properties:
        cleanup_complete:
          title: Cleanup Complete
          type: boolean
      required:
      - cleanup_complete
      title: DocumentCleanupResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    SkillActorRequest:
      properties:
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - workspace_id
      - user_id
      title: SkillActorRequest
      type: object
    SkillAuthoringResponse:
      properties:
        allowed_tools:
          items:
            enum:
            - list_root_items
            - list_folder_children
            - search_hierarchy
            - get_breadcrumb
            - get_document_metadata
            - get_document_content
            - create_folder
            - rename_folder
            - move_folder
            - move_document
            - rename_document
            - create_document
            - apply_document_edit
            type: string
          title: Allowed Tools
          type: array
        capabilities:
          items:
            enum:
            - document-create
            - document-edit
            - folder-organize
            - template
            type: string
          title: Capabilities
          type: array
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        instructions_markdown:
          anyOf:
          - type: string
          - type: 'null'
          title: Instructions Markdown
        issues:
          items:
            additionalProperties: true
            type: object
          title: Issues
          type: array
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        question:
          anyOf:
          - type: string
          - type: 'null'
          title: Question
        scope_type:
          anyOf:
          - enum:
            - personal
            - team
            type: string
          - type: 'null'
          title: Scope Type
        skill_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Skill Id
        skill_markdown:
          anyOf:
          - type: string
          - type: 'null'
          title: Skill Markdown
        status:
          enum:
          - clarification_required
          - blocked
          - proposal_ready
          - published
          title: Status
          type: string
        version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Version Id
      required:
      - status
      title: SkillAuthoringResponse
      type: object
...
    SkillDefinitionRequest:
      properties:
        allowed_tools:
          items:
            enum:
            - list_root_items
            - list_folder_children
            - search_hierarchy
            - get_breadcrumb
            - get_document_metadata
            - get_document_content
            - create_folder
            - rename_folder
            - move_folder
            - move_document
            - rename_document
            - create_document
            - apply_document_edit
            type: string
          title: Allowed Tools
          type: array
        capabilities:
          items:
            enum:
            - document-create
            - document-edit
            - folder-organize
            - template
            type: string
          minItems: 1
          title: Capabilities
          type: array
        description:
          minLength: 1
          title: Description
          type: string
        instructions_markdown:
          minLength: 1
          title: Instructions Markdown
          type: string
        name:
          maxLength: 63
          minLength: 1
          pattern: ^[a-z0-9][a-z0-9-]{0,62}$
          title: Name
          type: string
        user_id:
          minLength: 1
          title: User Id
          type: string
      required:
      - user_id
      - name
      - description
      - instructions_markdown
      - capabilities
      title: SkillDefinitionRequest
      type: object
    SkillDraftProposalRequest:
      properties:
        excluded_literals:
          items:
            type: string
          title: Excluded Literals
          type: array
        model:
          minLength: 1
          title: Model
          type: string
        provider:
          minLength: 1
          title: Provider
          type: string
        scope_type:
          enum:
          - personal
          - team
          title: Scope Type
          type: string
        source_runs:
          items:
            $ref: '#/components/schemas/SkillDraftSourceRunRequest'
          minItems: 1
          title: Source Runs
          type: array
        user_directives:
          items:
            type: string
          title: User Directives
          type: array
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - provider
      - model
      - workspace_id
      - user_id
      - scope_type
      - source_runs
      title: SkillDraftProposalRequest
      type: object
    SkillDraftSourceOperationRequest:
      properties:
        reason:
          minLength: 1
          title: Reason
          type: string
        tool_name:
          enum:
          - list_root_items
          - list_folder_children
          - search_hierarchy
          - get_breadcrumb
          - get_document_metadata
          - get_document_content
          - create_folder
          - rename_folder
          - move_folder
          - move_document
          - rename_document
          - create_document
          - apply_document_edit
          title: Tool Name
          type: string
      required:
      - tool_name
      - reason
      title: SkillDraftSourceOperationRequest
      type: object
    SkillDraftSourceRunRequest:
      properties:
        plan_summary:
          minLength: 1
          title: Plan Summary
          type: string
        request_summary:
          minLength: 1
          title: Request Summary
          type: string
        run_id:
          minLength: 1
          title: Run Id
          type: string
        status:
          const: completed
          title: Status
          type: string
        successful_operations:
          items:
            $ref: '#/components/schemas/SkillDraftSourceOperationRequest'
          minItems: 1
          title: Successful Operations
          type: array
      required:
      - run_id
      - status
      - request_summary
      - plan_summary
      - successful_operations
      title: SkillDraftSourceRunRequest
      type: object
    SkillPreviewResponse:
      properties:
        has_blocked_issues:
          title: Has Blocked Issues
          type: boolean
        lint_result:
          additionalProperties: true
          title: Lint Result
          type: object
      required:
      - lint_result
      - has_blocked_issues
      title: SkillPreviewResponse
      type: object
    SkillResponse:
      properties:
        enabled_version:
          anyOf:
          - $ref: '#/components/schemas/SkillVersionResponse'
          - type: 'null'
        id:
          title: Id
          type: string
        latest_version:
          anyOf:
          - $ref: '#/components/schemas/SkillVersionResponse'
          - type: 'null'
        owner_user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner User Id
        scope_type:
          title: Scope Type
          type: string
        slug:
          title: Slug
          type: string
        status:
          title: Status
          type: string
        workspace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Workspace Id
      required:
      - id
      - workspace_id
      - scope_type
      - owner_user_id
      - slug
      - status
      - enabled_version
      - latest_version
      title: SkillResponse
      type: object
    SkillTaskRequest:
      properties:
        kind:
          enum:
          - skill_author
          - skill_publish
          - skill_update
          title: Kind
          type: string
        payload:
          additionalProperties: true
          title: Payload
          type: object
        run_id:
          maxLength: 120
          minLength: 1
          title: Run Id
          type: string
        skill_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Skill Id
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - run_id
      - kind
      - workspace_id
      - user_id
      - payload
      title: SkillTaskRequest
      type: object
    SkillVersionResponse:
      properties:
        allowed_tools:
          items:
            type: string
          title: Allowed Tools
          type: array
        capabilities:
          items:
            type: string
          title: Capabilities
          type: array
        description:
          title: Description
          type: string
        id:
          title: Id
          type: string
        instructions_markdown:
          title: Instructions Markdown
          type: string
        lint_result:
          additionalProperties: true
          title: Lint Result
          type: object
        name:
          title: Name
          type: string
        status:
          title: Status
          type: string
        version:
          title: Version
          type: integer
      required:
      - id
      - version
      - name
      - description
      - instructions_markdown
      - capabilities
      - allowed_tools
      - lint_result
      - status
      title: SkillVersionResponse
      type: object
...
    TaskStatusResponse:
      properties:
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        id:
          title: Id
          type: string
        status:
          title: Status
          type: string
      required:
      - id
      - status
      title: TaskStatusResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  title: Fruition Pipeline Lab API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /internal/ai/tasks/documents/{document_id}/cancel:
    post:
      operationId: cancel_document_tasks_internal_ai_tasks_documents__document_id__cancel_post
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelTaskRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentCleanupResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Cancel Document Tasks
      tags:
      - ai-task-cancellation
  /internal/ai/tasks/{run_id}:
    get:
      operationId: task_status_internal_ai_tasks__run_id__get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: query
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      - in: query
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Task Status
      tags:
      - ai-task-cancellation
  /internal/ai/tasks/{run_id}/cancel:
    post:
      operationId: cancel_task_internal_ai_tasks__run_id__cancel_post
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelTaskRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Cancel Task
      tags:
      - ai-task-cancellation
  /internal/ai/tasks/{run_id}/rollback-backend:
    post:
      operationId: rollback_backend_task_internal_ai_tasks__run_id__rollback_backend_post
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelTaskRequest'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Rollback Backend Task
      tags:
      - ai-task-cancellation
...
  /skills:
    get:
      operationId: list_skills_skills_get
      parameters:
      - in: query
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      - in: query
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      - in: header
        name: X-Agent-Service-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Service-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SkillResponse'
                title: Response List Skills Skills Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Skills
      tags:
      - skills
  /skills/draft-from-runs/preview:
    post:
      operationId: propose_skill_draft_skills_draft_from_runs_preview_post
      parameters:
      - in: header
        name: X-Agent-Service-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Service-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillDraftProposalRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillAuthoringResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Propose Skill Draft
      tags:
      - skills
  /skills/preview:
    post:
      operationId: preview_skill_skills_preview_post
      parameters:
      - in: header
        name: X-Agent-Service-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Service-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillDefinitionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillPreviewResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Preview Skill
      tags:
      - skills
  /skills/tasks:
    post:
      operationId: execute_skill_task_skills_tasks_post
      parameters:
      - in: header
        name: X-Agent-Service-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Service-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillTaskRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Execute Skill Task
      tags:
      - skills
  /skills/{skill_id}:
    get:
      operationId: get_skill_skills__skill_id__get
      parameters:
      - in: path
        name: skill_id
        required: true
        schema:
          title: Skill Id
          type: string
      - in: query
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      - in: query
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      - in: header
        name: X-Agent-Service-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Service-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Skill
      tags:
      - skills
  /skills/{skill_id}/disable:
    post:
      operationId: disable_skill_skills__skill_id__disable_post
      parameters:
      - in: path
        name: skill_id
        required: true
        schema:
          title: Skill Id
          type: string
      - in: header
        name: X-Agent-Service-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Service-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillActorRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Disable Skill
      tags:
      - skills
  /skills/{skill_id}/enable:
    post:
      operationId: enable_skill_skills__skill_id__enable_post
      parameters:
      - in: path
        name: skill_id
        required: true
        schema:
          title: Skill Id
          type: string
      - in: header
        name: X-Agent-Service-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Service-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillActorRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Enable Skill
      tags:
      - skills
...